Arbitrary aspect ratio, using float instead of double, ver 1.81

+++ src/rasterizerC/color.h
@@ -5,16 +5,16 @@
 /*
 class Color
 {
-  double rgb[3];
-  double a;
+  float rgb[3];
+  float a;
 public:
 	// if this pixel has been drawn
 	//bool transparent;
 public:
-  Color(double r = 0.0, double g = 0.0, double b = 0.0, double a = 1.0);
+  Color(float r = 0.0, float g = 0.0, float b = 0.0, float a = 1.0);
   void draw(const Color &o);
-	void fill(double r, double g, double b);
-  Color fainter(double k);
+	void fill(float r, float g, float b);
+  Color fainter(float k);
   std::string as_ppm();
 	unsigned char* as_ppm(unsigned char *rgbbuf);
   // __repr__ operator<<
@@ -25,15 +25,15 @@ public:
 struct Color
 {
 	// {r, g, b} from 0.0 to 1.0
-	double rgb[3];
+	float rgb[3];
 	// alpha value, from 0.0 to 1.0
-	double a;
+	float a;
 };
 
-void init_Color(Color *color, double r = 0.0, double g = 0.0, double b = 0.0, double a = 1.0);
-Color init_Color(double r = 0.0, double g = 0.0, double b = 0.0, double a = 1.0);
+void init_Color(Color *color, float r = 0.0, float g = 0.0, float b = 0.0, float a = 1.0);
+Color init_Color(float r = 0.0, float g = 0.0, float b = 0.0, float a = 1.0);
 void draw_Color(Color *dst, const Color *src);
-Color faint_Color(const Color *color, double k);
+Color faint_Color(const Color *color, float k);
 unsigned char *as_ppm(const Color *color, unsigned char *rgbbuf);
 
 #endif //__COLOR_H__
