+++ src/rasterizerC/shape.h
@@ -22,9 +22,9 @@ protected:
 public:
   Shape(const Color *o_color = nullptr);
   virtual bool contains(const Vector &p) const = 0;
-  virtual double signed_distance_bound(const Vector &p) const = 0;
+  virtual float signed_distance_bound(const Vector &p) const = 0;
   void draw(PPMImage &image, int super_sampling);
-	virtual void get_parameters(std::vector<double> &paras, ShapeType *shapeType) = 0;
+	virtual void get_parameters(std::vector<float> &paras, ShapeType *shapeType) = 0;
 	// TODO: transform return true caller (derived) class obj
 	//template <class Shape_drv> // derived class of Shape
 	//virtual Shape_drv transform(const Transform &transform) const = 0;
